Add optional info about pci slot path for wireless
On Turris OS 6.0+ the default identification of wireless devices in
uci config became the `option path`.

Previously the `option macaddr` was used by default.

For example:

config wifi-device 'radio0'
    option path 'soc/soc:pcie/pci0000:00/0000:00:02.0/0000:02:00.0'

Due to the processing in foris-controller - matching the detected
hardware and current uci config to determine active and configured
wireless network interfaces - it is also needed to include the pci slot
path from /sys to be able to detect wireless interfaces correctly.

Closes: #21

Co-authored-by: Filip Hron <filip.hron@nic.cz>